refactor(graphql): extracted builder interface in separate file
